requirements.txt
ansible==2.7.0
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
Django==1.11
djangorestframework==3.8.2
idna==2.7
Jinja2==2.10
MarkupSafe==1.0
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
pytz==2018.5
PyYAML==3.13
six==1.11.0
